# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AssignmentInfo {
  title: string;
  studentName: string;
  rollNumber: string;
  courseName: string;
  instructorName: string;
  semester?: string;
  section?: string;
  date?: string;
}

export interface SourceFile {
  id: string;
  name: string;
  content: string;
  /** @minimum 0 */
  size: number;
}

export type GroupedFileConfidence = typeof GroupedFileConfidence[keyof typeof GroupedFileConfidence];


export const GroupedFileConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type GroupedFile = SourceFile & ({
  /** @nullable */
  detectedQuestion: number | null;
  detectedType: string;
  confidence: GroupedFileConfidence;
});

export interface AssignmentAnalysisInput {
  assignment: AssignmentInfo;
  /** @minItems 1 */
  files: SourceFile[];
}

export interface QuestionGroup {
  /** @minimum 1 */
  questionNumber: number;
  files: GroupedFile[];
}

export interface AssignmentAnalysis {
  assignment: AssignmentInfo;
  groups: QuestionGroup[];
  ungrouped: GroupedFile[];
  supportedExtensions: string[];
}

export type PdfSettingsPageSize = typeof PdfSettingsPageSize[keyof typeof PdfSettingsPageSize];


export const PdfSettingsPageSize = {
  A4: 'A4',
  Letter: 'Letter',
} as const;

export type PdfSettingsFontSize = typeof PdfSettingsFontSize[keyof typeof PdfSettingsFontSize];


export const PdfSettingsFontSize = {
  small: 'small',
  medium: 'medium',
  large: 'large',
} as const;

export interface PdfSettings {
  pageSize: PdfSettingsPageSize;
  fontSize: PdfSettingsFontSize;
  lineNumbers: boolean;
  syntaxHighlighting: boolean;
  coverPage: boolean;
}

export interface AssignmentPdfInput {
  assignment: AssignmentInfo;
  /** @minItems 1 */
  groups: QuestionGroup[];
  settings?: PdfSettings;
}