# lib/api-zod/src/generated/types/assignmentAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentInfo } from './assignmentInfo';
import type { GroupedFile } from './groupedFile';
import type { QuestionGroup } from './questionGroup';

export interface AssignmentAnalysis {
  assignment: AssignmentInfo;
  groups: QuestionGroup[];
  ungrouped: GroupedFile[];
  supportedExtensions: string[];
}