# lib/api-zod/src/generated/types/assignmentInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentInfo {
  title: string;
  studentName: string;
  rollNumber: string;
  courseName: string;
  instructorName: string;
  semester?: string;
  section?: string;
  date?: string;
}