# lib/api-zod/src/generated/types/assignmentPdfInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentInfo } from './assignmentInfo';
import type { PdfSettings } from './pdfSettings';
import type { QuestionGroup } from './questionGroup';

export interface AssignmentPdfInput {
  assignment: AssignmentInfo;
  /** @minItems 1 */
  groups: QuestionGroup[];
  settings?: PdfSettings;
}