# lib/api-zod/src/generated/types/groupedFile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroupedFileConfidence } from './groupedFileConfidence';
import type { SourceFile } from './sourceFile';

export type GroupedFile = SourceFile & ({
  /** @nullable */
  detectedQuestion: number | null;
  detectedType: string;
  confidence: GroupedFileConfidence;
});