# lib/api-zod/src/generated/types/pdfSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfSettingsFontSize } from './pdfSettingsFontSize';
import type { PdfSettingsPageSize } from './pdfSettingsPageSize';

export interface PdfSettings {
  pageSize: PdfSettingsPageSize;
  fontSize: PdfSettingsFontSize;
  lineNumbers: boolean;
  syntaxHighlighting: boolean;
  coverPage: boolean;
}