# lib/api-zod/src/generated/types/questionGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroupedFile } from './groupedFile';

export interface QuestionGroup {
  /** @minimum 1 */
  questionNumber: number;
  files: GroupedFile[];
}